+++ src/Pagination/PaginationArgs.php
@@ -36,7 +36,8 @@ class PaginationArgs
                 Cursor::decode($args),
             )
             // Handles cases "paginate" and "simple", which both take the same args.
-            : Arr::get($args, 'page', 1);
+            // Use default value 1 if page is not set or explicitly null
+            : (Arr::get($args, 'page') ?? 1);
 
         if ($first < 0) {
             throw new Error(self::requestedLessThanZeroItems($first));